did some refactor
