+++ src/components/sidebar/treeNode.js
@@ -1,55 +1,50 @@
 import React from "react";
-import styled from "react-emotion";
+import OpenedSvg from '../images/opened';
+import ClosedSvg from '../images/closed';
+import config from '../../../config';
 import Link from "../link";
 
-// eslint-disable-next-line no-unused-vars
-const TreeNode = styled(({className, active, level, ...props}) => {
-  if (level === 0) {
-    return (
-      <li className={className}>
-        <Link {...props} />
-      </li>
-    );
-  } else if (level === 1) {
-    const customClass = active ? 'active' : '';
-    return (
-      <li className={'subLevel ' + customClass}>
-        <Link {...props} />
-      </li>
-    );
-  } else {
-    return (
-      <li className={className}>
-        <Link {...props} />
-      </li>
-    );
+const TreeNode = ({className = '', setCollapsed, collapsed, url, title, items, ...rest}) => {
+  const isCollapsed = collapsed[url];
+  const collapse = () => {
+    setCollapsed(url);
   }
-})`
-  list-style: none;
+  const hasChildren = items.length !== 0;
+  const active =
+    location && (location.pathname === url || location.pathname === (config.gatsby.pathPrefix + url));
+  const calculatedClassName = `${className} item ${active ? 'active' : ''}`;
+  return (
+    <li
+      className={calculatedClassName}
+    >
+      {title && (
+        <Link
+          to={url}
+        >
+          {title}
+        </Link>)
+      }
 
-  a {
-    color: #fff;
-    text-decoration: none;
-    font-weight: ${({level}) => (level === 0 ? 700 : 400)};
-    padding: 0.45rem 0 0.45rem ${props => 2 + (props.level || 0) * 1}rem;
-    display: block;
-    position: relative;
-
-    &:hover {
-      background-color: #542683;
-    }
-
-    ${props =>
-    props.active &&
-    `
-      color: #fff;
-      background-color: #473485;
-    `} // external link icon
-    svg {
-      float: right;
-      margin-right: 1rem;
-    }
-  }
-`;
-
+      {!config.sidebar.frontLine && title && hasChildren ? (
+        <button
+          onClick={collapse}
+          className='collapser'>
+          {!isCollapsed ? <OpenedSvg /> : <ClosedSvg />}
+        </button>
+      ) : null}
+      {!isCollapsed && hasChildren ? (
+        <ul>
+          {items.map((item) => (
+            <TreeNode
+              key={item.url}
+              setCollapsed={setCollapsed}
+              collapsed={collapsed}
+              {...item}
+            />
+          ))}
+        </ul>
+      ) : null}
+    </li>
+  );
+}
 export default TreeNode
